Revert "Enable the option to combine each two registeres to hold IEEE-754 float representations"
This reverts commit dbdd8d33f631a8fc9221ca78143c9cb4ad9adefd.

+++ CommonControls/BaseForm.cs
@@ -16,7 +16,7 @@ namespace Modbus.Common
         private DisplayFormat _displayFormat = DisplayFormat.Integer;
         private CommunicationMode _communicationMode = CommunicationMode.TCP;
         protected Socket _socket;
-        protected readonly ushort[] _registerData;
+        protected readonly UInt16[] _registerData;
         private bool _logPaused = false;
 
         #region Form 
@@ -24,7 +24,7 @@ namespace Modbus.Common
         public BaseForm()
         {
             InitializeComponent();
-            _registerData = new ushort[65600];
+            _registerData = new UInt16[65600];
         }
         
         private void BaseFormLoading(object sender, EventArgs e)
@@ -678,14 +678,14 @@ namespace Modbus.Common
             if (tab.Text.Equals("...") && tabControl1.TabPages.Count < 20)
             {
                 DataTab dataTab = new DataTab();
-                dataTab.DataLength = 256;
+                dataTab.DataLength = ((ushort)(256));
                 dataTab.DisplayFormat = DisplayFormat.Integer;
                 dataTab.Location = new Point(3, 3);
                 dataTab.Name = "dataTab" + (tabControl1.TabPages.Count+1);
                 dataTab.RegisterData = _registerData;
                 dataTab.ShowDataLength = ShowDataLength;
                 dataTab.Size = new Size(839, 406);
-                dataTab.StartAddress = 0;
+                dataTab.StartAddress = ((ushort)(0));
                 dataTab.TabIndex = 0;
                 dataTab.OnApply += dataTab_OnApply;
                 TabPage tabPage = new TabPage();
